refactor: source changelog data directly from CHANGELOG.md
